+++ tests/server.py
@@ -2,13 +2,10 @@ from flask import Flask
 
 app = Flask(__name__)
 
-with open("tests/external.css") as fd:
-    STYLESHEET = fd.read()
-
 
 @app.route("/external.css")
 def stylesheet():
-    return STYLESHEET
+    return "h1 { color: blue; }"
 
 
 if __name__ == "__main__":