+++ python/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Skip selectors, that can't be parsed.
+- Validate `base_url` to be a valid URL.
 
 ### Fixed